% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyLift.R
\docType{package}
\name{easyLift}
\alias{easyLift}
\alias{easyLift-package}
\title{easyLift: Efficient parallelized compression of Hi-C Data}
\description{
The easyLift package contains functions to quickly convert
between different builds of the mouse and human genome. 
Note: this doens't support cross-species crossover.
}
