% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{pickOne}
\alias{pickOne}
\title{Pick one GRange from many in a list}
\usage{
pickOne(grl, returnBooVec = FALSE)
}
\arguments{
\item{grl}{A GRangeslist object to be picked on}

\item{returnBooVec}{If true, returns an S3
object where the first element is the flattened
GRanges object and the second is a boolean vector
indicating whether the region was successfully mapped or not.}
}
\value{
A flattened GRanges object of the same length
}
\description{
Function to pick a single GRanges object per
many in a given GRangeslist If none map, then a 
return a "chrZ" seqname is returned.
}
\examples{
from <- paste0(system.file('extdata',package='easyLift'),'/exbed/hg19.quick.bed')
#easyLiftOver(sf, map = "hg19_hg38") # spits out the lifted over bed file
liftedGRanges <- easyLiftOver(from, map = "hg19_hg38", pickOne = TRUE)

}
