% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easylift.R
\name{easylift}
\alias{easylift}
\title{Lift genomic coordinates from one genome assembly to another.}
\usage{
easylift(x, to, chain)
}
\arguments{
\item{x}{A GRanges object with genomic coordinates in the original assembly.}

\item{to}{The target genome assembly (e.g., "hg38").}

\item{chain}{The path to the chain file containing the liftover mapping.
Can be provided in gzipped or non-gzipped format.}
}
\value{
A GRanges object with lifted genomic coordinates.
}
\description{
This function takes a GRanges object with genomic coordinates in one genome
assembly and lifts them to another genome assembly using a chain file.
}
\examples{
# Lift over the coordinates of the first 10 genes in the hg19 assembly
# to the hg38 assembly
library(GenomicRanges)
library(easylift)
gr <- GRanges(seqname = Rle(c("chr1", "chr2"), c(100000, 100000)), 
ranges = IRanges(start = 1, end = 200000))
genome(gr) <- "hg19"
chain <- system.file("extdata/hg19ToHg38.over.chain.gz", package="easylift")
easylift(gr, "hg38", chain)
}
