% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easylift.R
\name{easylift}
\alias{easylift}
\title{Lift genomic coordinates from one genome assembly to another.}
\usage{
easylift(x, to, chain)
}
\arguments{
\item{x}{A GRanges object with genomic coordinates in the original assembly.}

\item{to}{The target genome assembly (e.g., "hg38").}

\item{chain}{The path to the chain file containing the liftover mapping.
Can be provided in gzipped or non-gzipped format.}
}
\value{
A GRanges object with lifted genomic coordinates.
}
\description{
This function takes a GRanges object with genomic coordinates in one genome
assembly and lifts them to another genome assembly using a chain file.
}
