% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{easyLiftOver}
\alias{easyLiftOver}
\title{easyLiftOver}
\usage{
easyLiftOver(from, map = "hg19_hg38", to = NULL)
}
\arguments{
\item{from}{A .bed file or GRanges object containing
coordinates to be lifted over.}

\item{map}{Reference genome mapping logic {FROM}_{TO}
or alternatively, the filepath associated with a valid
.chain file.}

\item{to}{File to write liftover output to. By default,
we just append "over" to the input file name. This 
parameter is ignored if the user specifies a GRanges object.}
}
\value{
A GRanges if "from" is a GRanges. 
Otherwise, it will write a file with "over" appeneded
with the lifted over coordinates
}
\description{
Function to get permuted data while maintaining biases
}
\examples{
from <- paste0(system.file('extdata',package='easyLift'),'/exbed/hg19.quick.bed')
#easyLiftOver(sf, map = "hg19_hg38") # spits out the lifted over bed file
liftedGRanges <- easyLiftOver(from, map = "hg19_hg38")

}
