% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{easyLiftOver}
\alias{easyLiftOver}
\title{easyLiftOver}
\usage{
easyLiftOver(from, to = NULL, map = "hg19_hg38", write = TRUE)
}
\arguments{
\item{from}{.bed file}

\item{to}{.bed file of the output after lifting over}

\item{map}{Reference genome mapping logic {FROM}_{TO}}

\item{write}{Either write (by default) to file if TRUE 
or return a GRanges object of the result.}
}
\value{
A GRanges if write == FALSE; spits out a new
bed file if write == TRUE.
}
\description{
Function to get permuted data while maintaining biases
}
\examples{
sf <- paste0(system.file('extdata',package='easyLift'),'/exbed/hg19.quick.bed')
#easyLiftOver(sf, map = "hg19_hg38") # spits out the lifted over bed file
liftedGRanges <- easyLiftOver(sf, map = "hg19_hg38", write = FALSE)

}
